+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Enhancements
 
 * Add `A.Vector.flat_map/2`
+* Add `A.Vector.reverse/2` and `A.Enum.reverse/2`
 * `A.OrdMap.drop/2` will only try to period rebuild once after dropping all keys
 
 ### Bug fixes